--- modules/aggs-composite/src/main/java/org/elasticsearch/search/aggregations/composite/spi/CompositeNamedXContentProvider.java
@@ -1,42 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.search.aggregations.composite.spi;
-
-import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.xcontent.ContextParser;
-import org.elasticsearch.common.xcontent.NamedXContentRegistry;
-import org.elasticsearch.plugins.spi.NamedXContentProvider;
-import org.elasticsearch.search.aggregations.Aggregation;
-import org.elasticsearch.search.aggregations.composite.CompositeAggregationBuilder;
-import org.elasticsearch.search.aggregations.composite.ParsedComposite;
-
-import java.util.List;
-
-import static java.util.Collections.singletonList;
-
-public class CompositeNamedXContentProvider implements NamedXContentProvider {
-
-    @Override
-    public List<NamedXContentRegistry.Entry> getNamedXContentParsers() {
-        ParseField parseField = new ParseField(CompositeAggregationBuilder.NAME);
-        ContextParser<Object, Aggregation> contextParser = (p, name) -> ParsedComposite.fromXContent(p, (String) name);
-        return singletonList(new NamedXContentRegistry.Entry(Aggregation.class, parseField, contextParser));
-    }
-}
