--- modules/aggs-composite/src/test/resources/rest-api-spec/test/composite/10_basic.yml
@@ -1,13 +1,0 @@
-# Integration tests for Composite aggs plugin
-#
-"Composite aggs loaded":
-    - do:
-        cluster.state: {}
-
-    # Get master node id
-    - set: { master_node: master }
-
-    - do:
-        nodes.info: {}
-
-    - match:  { nodes.$master.modules.0.name: aggs-composite }
