--- modules/aggs-composite/src/test/resources/rest-api-spec/test/composite/20_composite.yml
@@ -1,196 +1,0 @@
----
-setup:
-  - do:
-        indices.create:
-          index: test
-          body:
-              mappings:
-                doc:
-                  properties:
-                    keyword:
-                      type: keyword
-                    long:
-                      type: long
-
-  - do:
-      index:
-        index: test
-        type:  doc
-        id:    1
-        body:  { "keyword": "foo", "long": [10, 20] }
-
-  - do:
-      index:
-        index: test
-        type:  doc
-        id:    2
-        body:  { "keyword": ["foo", "bar"] }
-
-  - do:
-      index:
-        index: test
-        type:  doc
-        id:    3
-        body:  { "keyword": "bar", "long": [100, 0] }
-
-  - do:
-      index:
-        index: test
-        type:  doc
-        id:    4
-        body:  { "keyword": "bar", "long": [1000, 0] }
-
-  - do:
-      indices.refresh:
-        index: [test]
-
----
-"Simple Composite aggregation":
-  - skip:
-      version: " - 6.99.99"
-      reason:  this uses a new API that has been added in 7.0
-
-
-  - do:
-      search:
-        index: test
-        body:
-          aggregations:
-            test:
-              composite:
-                sources: [
-                  "kw": {
-                    "terms": {
-                      "field": "keyword"
-                    }
-                  }
-                ]
-
-  - match: {hits.total: 4}
-  - length: { aggregations.test.buckets: 2 }
-  - match: { aggregations.test.buckets.0.key.kw: "bar" }
-  - match: { aggregations.test.buckets.0.doc_count: 3 }
-  - match: { aggregations.test.buckets.1.key.kw: "foo" }
-  - match: { aggregations.test.buckets.1.doc_count: 2 }
-
----
-"Nested Composite aggregation":
-  - skip:
-      version: " - 6.99.99"
-      reason:  this uses a new API that has been added in 7.0
-
-
-  - do:
-      search:
-        index: test
-        body:
-          aggregations:
-            test:
-              composite:
-                sources: [
-                  {
-                    "long": {
-                      "terms": {
-                        "field": "long"
-                      }
-                    }
-                  },
-                  {
-                    "kw": {
-                      "terms": {
-                        "field": "keyword"
-                      }
-                    }
-                  }
-                ]
-
-  - match: {hits.total: 4}
-  - length: { aggregations.test.buckets: 5 }
-  - match: { aggregations.test.buckets.0.key.long: 0}
-  - match: { aggregations.test.buckets.0.key.kw: "bar" }
-  - match: { aggregations.test.buckets.0.doc_count: 2 }
-  - match: { aggregations.test.buckets.1.key.long: 10 }
-  - match: { aggregations.test.buckets.1.key.kw: "foo"}
-  - match: { aggregations.test.buckets.1.doc_count: 1 }
-  - match: { aggregations.test.buckets.2.key.long: 20 }
-  - match: { aggregations.test.buckets.2.key.kw: "foo" }
-  - match: { aggregations.test.buckets.2.doc_count: 1 }
-  - match: { aggregations.test.buckets.3.key.long: 100}
-  - match: { aggregations.test.buckets.3.key.kw: "bar" }
-  - match: { aggregations.test.buckets.3.doc_count: 1 }
-  - match: { aggregations.test.buckets.4.key.long: 1000 }
-  - match: { aggregations.test.buckets.4.key.kw: "bar" }
-  - match: { aggregations.test.buckets.4.doc_count: 1 }
-
----
-"Aggregate After":
-  - skip:
-      version: " - 6.99.99"
-      reason:  this uses a new API that has been added in 7.0
-
-
-  - do:
-      search:
-        index: test
-        body:
-          aggregations:
-            test:
-              composite:
-                sources: [
-                  {
-                    "long": {
-                      "terms": {
-                        "field": "long"
-                      }
-                    }
-                  },
-                  {
-                    "kw": {
-                      "terms": {
-                        "field": "keyword"
-                      }
-                    }
-                  }
-                ]
-                after: { "long": 20, "kw": "foo" }
-
-  - match: {hits.total: 4}
-  - length: { aggregations.test.buckets: 2 }
-  - match: { aggregations.test.buckets.0.key.long: 100 }
-  - match: { aggregations.test.buckets.0.key.kw: "bar" }
-  - match: { aggregations.test.buckets.0.doc_count: 1 }
-  - match: { aggregations.test.buckets.1.key.long: 1000 }
-  - match: { aggregations.test.buckets.1.key.kw: "bar" }
-  - match: { aggregations.test.buckets.1.doc_count: 1 }
-
----
-"Invalid Composite aggregation":
-  - skip:
-      version: " - 6.99.99"
-      reason:  this uses a new API that has been added in 7.0
-
-  - do:
-      catch:      /\[composite\] aggregation cannot be used with a parent aggregation/
-      search:
-        index: test
-        body:
-          aggregations:
-            test:
-              terms:
-                field: long
-              aggs:
-                nested:
-                  composite:
-                    sources: [
-                      {
-                        "kw": {
-                          "terms": {
-                            "field": "keyword"
-                          }
-                        }
-                      }
-                    ]
-
-
-
-
