+++ settings.gradle
@@ -31,7 +31,6 @@ List projects = [
   'test:fixtures:old-elasticsearch',
   'test:logger-usage',
   'modules:aggs-matrix-stats',
-  'modules:aggs-composite',
   'modules:analysis-common',
   'modules:ingest-common',
   'modules:lang-expression',
